Convert task classes to more generic so they can be re-used
