+++ cyclops/tasks/__init__.py
@@ -4,5 +4,3 @@ from cyclops.tasks.classification import (
     BinaryTabularClassificationTask,
     MultilabelImageClassificationTask,
 )
-from cyclops.tasks.cxr_classification import CXRClassificationTask
-from cyclops.tasks.mortality_prediction import MortalityPredictionTask
